refactor(prefs): extract Notch and Misc tabs from General preferences
Move notch overlay toggle to a dedicated Notch tab and WPM/keystroke
overlay toggles to a new Misc tab. Add a new "Features" group in the
preferences sidebar for both. General tab now only contains core app
settings (reopen workspace, new tab/worktree defaults, shell info).
