+++ src/renderer/src/components/preferences/PreferencesModal.svelte
@@ -15,6 +15,8 @@
   import TaskTrackerPrefs from './TaskTrackerPrefs.svelte'
   import PrivacyPrefs from './PrivacyPrefs.svelte'
   import FileWatcherPrefs from './FileWatcherPrefs.svelte'
+  import NotchPrefs from './NotchPrefs.svelte'
+  import MiscPrefs from './MiscPrefs.svelte'
 
   let { section: initialSection }: { section?: string } = $props()
 
@@ -22,6 +24,7 @@
 
   const groups = [
     { label: 'General', sections: ['General', 'Updates', 'Privacy', 'Shortcuts'] },
+    { label: 'Features', sections: ['Notch', 'Misc'] },
     { label: 'Appearance', sections: ['Appearance', 'Sidebar'] },
     { label: 'AI Agents', sections: ['Claude', 'Gemini'] },
     { label: 'Dev Tools', sections: ['Terminal', 'Tools', 'Git', 'Tasks', 'File Watcher'] },
@@ -113,6 +116,10 @@
         <PrivacyPrefs />
       {:else if activeSection === 'Shortcuts'}
         <ShortcutsPrefs />
+      {:else if activeSection === 'Notch'}
+        <NotchPrefs />
+      {:else if activeSection === 'Misc'}
+        <MiscPrefs />
       {/if}
     </div>
   </div>
